portfolio_project/backend/requirements.txt
Django==5.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.4.0
Pillow==10.4.0
SQLAlchemy==2.0.35
PyMySQL==1.1.0
cryptography==41.0.7
python-dotenv==1.0.0
werkzeug==3.0.1